% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFactorialHMNL-package.R
\docType{package}
\name{bhmnl-package}
\alias{bhmnl}
\alias{bhmnl-package}
\title{bhmnl: Bayesian hierarichical multinomil logistic with possible factorial responses}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tnearey/bhmnl}
  \item Report bugs at \url{https://github.com/tnearey/bhmnl/issues}
}

}
\author{
\strong{Maintainer}: Terrance Nearey \email{noone@nowhere.net} (YOUR-ORCID-ID)

}
\keyword{internal}
