% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpandSumToZeroCoefs.R
\name{buildFamilyMainEffectsList}
\alias{buildFamilyMainEffectsList}
\title{Build a  main effects list for a family of coefficient names}
\usage{
buildFamilyMainEffectsList(thisFamilysCoefs)
}
\arguments{
\item{thisFamilysCoefs}{coefficients, usually for a single family (or for entire set of families for a formula)}
}
\value{
mainEffectsList  named list of main effects with levels and attributes nfac gfacType and trueFactorNames
}
\description{
Build a  main effects list for a family of coefficient names
}
\examples{
expandedCoefs=c("C|b",  "C|d",  "V|a",  "V|i",
  "V|u",  "C|b:V|a",   "C|d:V|a",  "C|b:V|i",  "C|d:V|i",  "C|b:V|u",
   "C|d:V|u",  "C|b:F1", "C|d:F1",  "V|a:F2",  "V|i:F2",  "V|u:F2",
    "C|b:cond|1",  "C|d:cond|1","C|b:cond|2",  "C|d:cond|2",
     "C|b:cond|3",  "C|d:cond|3")
thisMainEffectsList=buildFamilyMainEffectsList(expandedCoefs)
print(str(thisMainEffectsList))
}
