% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{catln}
\alias{catln}
\title{Call `cat()` with trailing newline}
\usage{
catln(...)
}
\arguments{
\item{...}{args to pass to cat()}
}
\description{
Call `cat()` with trailing newline
}
\examples{
catln('Boo',27,'hoo')
}
