% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpandSumToZeroCoefs.R
\name{checkMarginalsSumToZero}
\alias{checkMarginalsSumToZero}
\title{Check marginal sum-to-zero constraints of expanded coefficients}
\usage{
checkMarginalsSumToZero(expandedCoefs)
}
\arguments{
\item{expandedCoefs}{a  [K x 1] matrix of named {fac|lev:} coefficients (expanded) that obey marginal sum-to-zero}
}
\value{
logical (TRUE if all sum to zero constraints are met)
}
\description{
Check marginal sum-to-zero constraints of expanded coefficients
}
\examples{
print ('No examples see vignett TestExpandSumToZeroCoefs.Rmd')
}
