% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpandSumToZeroCoefs.R
\name{constructFactorialDtb}
\alias{constructFactorialDtb}
\title{Construct a  factorial data table (facDtb)
A factorial data table (facDtb) is a   data table of main effect factor levels and filler covariates (1.0)  sufficient to cover the highest level  of factors and covariate interactions. "True factors" are identified by name in an attribute. Non-true factors are filler covariates. This is often called one family at a time.}
\usage{
constructFactorialDtb(mainEffectsList)
}
\arguments{
\item{mainEffectsList}{a main effects list as created by buildFamilyMainEffectsList() or a charvec of expanded coefficient names that can be made into one}
}
\description{
Construct a  factorial data table (facDtb)
A factorial data table (facDtb) is a   data table of main effect factor levels and filler covariates (1.0)  sufficient to cover the highest level  of factors and covariate interactions. "True factors" are identified by name in an attribute. Non-true factors are filler covariates. This is often called one family at a time.
}
\examples{
expandedCoefs=c("C|b",  "C|d",  "V|a",  "V|i",
  "V|u",  "C|b:V|a",   "C|d:V|a",  "C|b:V|i",  "C|d:V|i",  "C|b:V|u",
   "C|d:V|u",  "C|b:F1", "C|d:F1",  "V|a:F2",  "V|i:F2",  "V|u:F2",
    "C|b:cond|1",  "C|d:cond|1","C|b:cond|2",  "C|d:cond|2",
     "C|b:cond|3",  "C|d:cond|3")
famwiseDtb=constructFactorialDtb(expandedCoefs)
print(str(famwiseDtb))
}
