% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFactoredCoefNames.R
\name{createResidualMaker}
\alias{createResidualMaker}
\title{Create a residual maker matrix from coefficient names.}
\usage{
createResidualMaker(coefNames, shouldReturnBigMatrix = FALSE,
  showLMTest = FALSE)
}
\arguments{
\item{coefNames}{coefficient names a e.g. c("C|d", "V|a",... "C_d:V_a: ...., C|b:V|a:y")}

\item{shouldReturnBigMatrix}{default FALSE. If true, RM_bigMat component returned}

\item{showLMTest}{default FALSE. f True an example test is run for random coefficients.}
}
\value{
result: a list with components familyNames, RM_list and optionally RM_bigMat
}
\description{
Create a residual maker matrix from coefficient names.
}
\details{
Given an character vector of coefficient names for a HMNL, this function creates
   familyNames would include "C", "V", C:V", "C:V:y"
   result$RM_list a list of residual maker matrices for anihilating all marginal sums of
  factorial coefficients.
and result$RM_bigMat  a large (ncoef x ncoef) matrix with non-zero blocks constisting of
}
\examples{
coefNames=getTestCoefNames()
result=createResidualMaker(coefNames, shouldReturnBigMatrix=TRUE, showLMTest=TRUE)
if (requireNamespace("nutils", quietly=TRUE)) {
   nutils::showvars(result)
 }else{
   print(str(result))
 }
}
