% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFactoredCoefNames.R
\name{getFamilyWiseCoefList}
\alias{getFamilyWiseCoefList}
\title{Get the familynames for each coefficient and organize into list}
\usage{
getFamilyWiseCoefList(coefNames)
}
\value{
a named list (by family names) of coefficients
for each family
}
\description{
Get the familynames for each coefficient and organize into list
}
\examples{
fwCoefNames=getFamilyWiseCoefList(c("A|a, A|b, B|a, B|b, A|a:B|b"))
print(str(fwCoefNames))
}
\seealso{
getFamNamesFromCoefNames()
}
