% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFactoredCoefNames.R
\name{getFamNamesFromCoefNames}
\alias{getFamNamesFromCoefNames}
\title{Get family names from coefficient names (several coefNames may have same family name)}
\usage{
getFamNamesFromCoefNames(coefNames)
}
\arguments{
\item{coefNames}{a character vector of coefficient names}
}
\value{
a character vector of family names (of same length as input)
}
\description{
Get family names from coefficient names (several coefNames may have same family name)
}
\examples{
print(getFamNamesFromCoefNames(c('a|1:b_2:x','a_1','b_2', 'x')))
}
\seealso{
getFamilyWiseCoefList()
}
