% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFactoredCoefNames.R
\name{getGFacAndLevNames}
\alias{getGFacAndLevNames}
\title{getGFacAndLevNames
Get general factor and factor level names}
\usage{
getGFacAndLevNames(theseCoefs)
}
\arguments{
\item{theseCoefs}{a character vector of coefficient names of form <gfacName>("|"<levelName>)
separated by ":"}
}
\value{
returns a gfacLevList with components $fac and $lev
 $levels is NA for  covariate
}
\description{
getGFacAndLevNames
Get general factor and factor level names
}
\examples{
print('noexamples')
}
