% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFactoredCoefNames.R
\name{getOneFamilyDfrAndCFacStr}
\alias{getOneFamilyDfrAndCFacStr}
\title{getOneFamilyDfrAndCFacStr Get a data frame and a categorical factor string of a coefficient family name}
\usage{
getOneFamilyDfrAndCFacStr(thisFam, coefNames, gcoefFamCv)
}
\arguments{
\item{thisFam}{-- a coefficient family name  (string) egs C:V:F1, C,V}

\item{coefNames}{-- character vector of coefficient names C|b:F1 etc.}

\item{gcoefFamCv}{-- a character vector of coefFam names}
}
\value{
a list with components dfr - data frame reflecting true factor  and cFacStr a string with the "true" multi level factors involved (covariates  removed)
}
\description{
getOneFamilyDfrAndCFacStr Get a data frame and a categorical factor string of a coefficient family name
}
\details{
dataframe  and  cFacStr are useful useful for creating model matrix to
 remove the marginal values of factors mm = model.matrix.
}
\examples{
print('noexamples')
}
