% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFactoredCoefNames.R
\name{getTestCoefNames}
\alias{getTestCoefNames}
\title{getTestCoefNames}
\usage{
getTestCoefNames()
}
\value{
list of test coefficient names
}
\description{
getTestCoefNames
}
\examples{
print(' no examples')
}
