% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{logsumexp}
\alias{logsumexp}
\title{Log of sum of exp of vector (numerically stable.)}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{vector}
}
\value{
log(sum(exp(x))) but better
}
\description{
Log of sum of exp of vector (numerically stable.)
}
\examples{
 logsumexp(c(1,2,3))
}
