% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{make_catRespMMFormStr}
\alias{make_catRespMMFormStr}
\title{make a "safe" categorial response model.matrix formula string, forcing all : to *}
\usage{
make_catRespMMFormStr(formSpec, respVarColNames, stimVarColNames)
}
\arguments{
\item{formSpec}{a formula or formula}

\item{respVarColNames}{charvec of response variable column names (all should be factors)}

\item{stimVarColNames}{charvec of response variable column names (factors or covariates)}
}
\value{
returns a character string
}
\description{
make a "safe" categorial response model.matrix formula string, forcing all : to *
}
\details{
Currently this is not used.
}
\examples{
respVarColnames=c('syl','C','V')
stimVarColNames=c('F1','F2','cond')
formStr=make_catRespMMFormStr(formSpec="syl~C*V+C:F1+C:F2+C:cond")
print(formStr)

}
