% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{make_indicatorDmx}
\alias{make_indicatorDmx}
\title{make_indicatorDmx Make a Design Matrix (sr- or rs-)}
\usage{
make_indicatorDmx(form, respFacStimDtbPadded, respVarColNames,
  stimVarColNames)
}
\arguments{
\item{form}{a formula or formula string}

\item{respFacStimDtbPadded}{a padded factorial StimDtb}

\item{respVarColNames}{column names for response variavles}

\item{stimVarColNames}{column names for stimulus variables}
}
\value{
returns an rsDmx or an srDmx
}
\description{
make_indicatorDmx Make a Design Matrix (sr- or rs-)
}
\examples{
cat('no examples \\n')
}
