% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{make_rsIndicatorDmxFrom_srDtb}
\alias{make_rsIndicatorDmxFrom_srDtb}
\title{make_rsIndicatorDmxFrom_srDtb}
\usage{
make_rsIndicatorDmxFrom_srDtb(form = NULL, srDtb = NULL,
  stimVarColNames = NULL, respFacDescr = NULL,
  includeMasterCatInDesired = TRUE)
}
\arguments{
\item{form}{a formula or formula string with response factors and interactions of
response factors with stimulus parameters   e.g. ~ C*V+ C:F1+ V:F2}

\item{srDtb}{a stimulus(-response0 data table
(Note: trial-by-trial responses are required only for model fitting and
for compling respDtbs from full srDtb's .
stimulus only tables can be used in e.g. test data generation)}

\item{stimVarColNames}{names of stimulus factors}

\item{respFacDescr}{either a list of response factor names  (first of which must be the "master category")
or a respFacDtb (of which all response col names will be used )}

\item{includeMasterCatInDesired}{(defalt TRUE) include the master response category in
the rsDmx response factors}
}
\value{
a rsDmx: a response-stimulus data matrix with attributes 'masterCatLevels'
}
\description{
make_rsIndicatorDmxFrom_srDtb
}
\examples{
cat('no examples \\n')
}
