% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{make_stimRespFacDtb}
\alias{make_stimRespFacDtb}
\title{Make a stimulus-response-factor data.table from respFacDeb and srDtb}
\usage{
make_stimRespFacDtb(respFacDtb, srDtb, stimVarColNames)
}
\arguments{
\item{respFacDtb}{response factor Data table}

\item{srDtb}{stimulus response data table}

\item{stimVarColNames}{names for stimulus variables in srDtb to be retained}
}
\value{
returns a stimRespFacDtb
}
\description{
Make a stimulus-response-factor data.table from respFacDeb and srDtb
}
\examples{
cat('no examples \\n')
}
