% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{make_sumDmx}
\alias{make_sumDmx}
\title{Make a sum to zero Design Matrix (sr- or rs-)}
\usage{
make_sumDmx(form, respFacStimDtb, respVarColNames, stimVarColNames)
}
\arguments{
\item{form}{a formula or formula string}

\item{respVarColNames}{column names for response variavles}

\item{stimVarColNames}{column names for stimulus variables}

\item{respFacStimDtbPadded}{a padded factorial StimDtb}
}
\value{
returns an rsDmx or an srDmx
}
\description{
Make a sum to zero Design Matrix (sr- or rs-)
}
\examples{
cat('no examples \\n')
}
