% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpandSumToZeroCoefs.R
\name{makeFamilyWiseDmxFromFactorialDtb}
\alias{makeFamilyWiseDmxFromFactorialDtb}
\title{Make family-wise data matrix from a facDtb}
\usage{
makeFamilyWiseDmxFromFactorialDtb(facDtb)
}
\arguments{
\item{facDtb}{}
}
\value{
a  stimulus-response MNL-compatible model.matrix
}
\description{
Make family-wise data matrix from a facDtb
}
\examples{
print('no example')
}
