% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFactoredCoefNames.R
\name{makeModelMatFromFamilyDfrAndCFacStr}
\alias{makeModelMatFromFamilyDfrAndCFacStr}
\title{makeModelMatFromFamilyDfrAndCFacStr}
\usage{
makeModelMatFromFamilyDfrAndCFacStr(dfr, cFacStr)
}
\arguments{
\item{dfr}{data frame as created in getOneFamilyDfrAndCFacStr}

\item{cFacStr}{string with true facrors (getOneFamilyDfrAndCFacStr)}
}
\value{
model matrix appropriate for marginal-centering of the factorial coeefficient family
}
\description{
makeModelMatFromFamilyDfrAndCFacStr
}
\examples{
print('no example  makeModelMatFromFamilyDfrAndCFacStr')
}
