% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{myIndicatorContrFun}
\alias{myIndicatorContrFun}
\title{A decorated contrast function always forcing INDICATOR levels}
\usage{
myIndicatorContrFun(levLabs, base = 1, contrasts = TRUE)
}
\arguments{
\item{levLabs}{charvec of level labels}

\item{base}{base category}

\item{contrasts}{contrast=TRUE (ignored)}
}
\value{
contr returns contrast matrix
}
\description{
A decorated contrast function always forcing INDICATOR levels
}
\examples{
cat('no examples\\n')
}
