% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{mySumContrFun}
\alias{mySumContrFun}
\title{A decorated contrast function always forcing sum-to-zero  levels (last cat is ref)}
\usage{
mySumContrFun(levLabs, base = NULL, contrasts = TRUE)
}
\arguments{
\item{levLabs}{charvec of level labels}

\item{base}{( not used should be null)}

\item{contrasts}{contrast=TRUE (ignored)}
}
\value{
contr returns contrast matrix
}
\description{
A decorated contrast function always forcing sum-to-zero  levels (last cat is ref)
}
\examples{
cat('no examples\\n')
}
