% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napp2logistic.R
\name{napp2logistic}
\alias{napp2logistic}
\title{Convert a (equi-covariance) NAPP model to logistic regression coefficients}
\usage{
napp2logistic(meanColVecs, covMat, priors = NULL)
}
\arguments{
\item{meanColVecs}{[nVar x nGroup] matrix of  means}

\item{covMat}{[nVar x nVar] common covariance matrix}

\item{priors}{[nGroup x 1] vector of priors}
}
\value{
bMat [(nVar+1) x 1] vector of coefficients s.t. xmatAug %*% b yields logistic scores
 and xmatAug rbind(1, xmat)  where xmat is rbind(xvar1,xvar2,xvar3...)
}
\description{
Convert a (equi-covariance) NAPP model to logistic regression coefficients
}
\examples{
  mu=matrix(c(-2, 0, 2,
-2, 0, -2),2,3, byrow = FALSE);
covMat=matrix(c(1, .5,
               .5, 1), 2,2, byrow=FALSE);
bMat=napp2logistic(meanColVecs,covMat))
print(bMat)
}
