% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{normvec}
\alias{normvec}
\title{Normalize  elements of(hopefully all positive ) vector (max forced to 1.0)}
\usage{
normvec(x)
}
\arguments{
\item{x}{vector of non-negatives}
}
\value{
normalized vector (sums to 1s)
}
\description{
Normalize  elements of(hopefully all positive ) vector (max forced to 1.0)
}
\examples{
normvec(c(1,2,3))
}
