% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{padLevels_facStim_srDtb}
\alias{padLevels_facStim_srDtb}
\title{Pad the levels of all factors of an srDtb data table.}
\usage{
padLevels_facStim_srDtb(srDtb, stimVarColNames)
}
\arguments{
\item{srDtb}{a (trial-wise) data table with stimulus factors that
may need "_" level padding}

\item{stimVarColNames}{names of columns including stimuli}
}
\value{
the srDtb with padded stimulus factor levels
}
\description{
Pad the levels of all factors of an srDtb data table.
}
\examples{
print("no example")
}
