% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneStanfitObj.R
\name{pruneStanfitObj}
\alias{pruneStanfitObj}
\title{Trim a stanfit object to only wanted names}
\usage{
pruneStanfitObj(object, pars)
}
\arguments{
\item{object}{stanfit object to prune}

\item{pars}{a charvec of parameter names to retain (lp_ is always included)}
}
\value{
returns a trimmed stanfit object
}
\description{
Trim a stanfit object to only wanted names
}
\examples{
print('No example')
}
