% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialResponseLogisticExpandedDesign.R
\name{softmax}
\alias{softmax}
\title{Softmax function exp(x)/logsumexp(x)}
\usage{
softmax(x)
}
\arguments{
\item{x}{vector (of eg log likelihoods)}
}
\value{
exp(x)/logsumexp(x)
}
\description{
Softmax function exp(x)/logsumexp(x)
}
\examples{
softmax(log(c(1,2,3)))
}
